ICU-23005 Replace use of CharString with Locale to simplify the code.
This brings a lot of the code that uses LocaleBased back to how it was
before the refactoring of ICU-23000, accessing the validLocale and
actualLocale fields directly instead of through LocaleBased helper
functions that now no longer are necessary.

This also makes it possible to remove the U_LOCALE_BASED() macro and
LocaleBased then ends up becoming a trivial class that now contains
nothing more than two static helper functions.